Move SelectOption
