+++ packages/twenty-front/src/modules/object-record/record-field/form-types/components/FormCountryCodeSelectInput.tsx
@@ -3,9 +3,9 @@ import { IconCircleOff, IconComponentProps } from 'twenty-ui';
 
 import { FormSelectFieldInput } from '@/object-record/record-field/form-types/components/FormSelectFieldInput';
 import { VariablePickerComponent } from '@/object-record/record-field/form-types/types/VariablePickerComponent';
-import { SelectOption } from '@/spreadsheet-import/types';
 import { useCountries } from '@/ui/input/components/internal/hooks/useCountries';
 import { CountryCode } from 'libphonenumber-js';
+import { SelectOption } from '~/types/SelectOption';
 
 export type FormCountryCodeSelectInputUpdatedValue = CountryCode | '';
