+++ packages/twenty-front/src/modules/object-record/record-field/form-types/components/FormCountrySelectInput.tsx
@@ -3,9 +3,8 @@ import { IconCircleOff, IconComponentProps } from 'twenty-ui';
 
 import { FormSelectFieldInput } from '@/object-record/record-field/form-types/components/FormSelectFieldInput';
 import { VariablePickerComponent } from '@/object-record/record-field/form-types/types/VariablePickerComponent';
-import { SelectOption } from '@/spreadsheet-import/types';
 import { useCountries } from '@/ui/input/components/internal/hooks/useCountries';
-
+import { SelectOption } from '~/types/SelectOption';
 export const FormCountrySelectInput = ({
   selectedCountryName,
   onChange,
