+++ packages/twenty-front/src/modules/ui/field/input/components/SelectInput.tsx
@@ -1,6 +1,6 @@
-import { SelectOption } from '@/spreadsheet-import/types';
 import { SelectInput as SelectBaseInput } from '@/ui/input/components/SelectInput';
 import { SelectableList } from '@/ui/layout/selectable-list/components/SelectableList';
+import { SelectOption } from '~/types/SelectOption';
 
 type SelectInputProps = {
   selectableListComponentInstanceId: string;
